+++ Netstroker/Server.cpp
@@ -48,7 +48,7 @@ namespace Netstroker
 
 	void Server::Frame()
 	{
-		for (int i = 0; i < connections.size(); i++)
+		for (unsigned int i = 0; i < connections.size(); i++)
 		{
 			if (connections[i].pm_outgoing.HasPendingPackets())
 			{
@@ -185,7 +185,7 @@ namespace Netstroker
 					{
 						if (pm.currentTask == PacketManagerTask::ProcessPacketSize) // sned packet size
 						{
-							pm.currentPacketSize = pm.Retrieve()->buffer.size();
+							pm.currentPacketSize = (uint16_t)pm.Retrieve()->buffer.size();
 							uint16_t bigEndianPacketSize = htons(pm.currentPacketSize);
 							int bytesSent = send(use_fd[i].fd, (char*)(&bigEndianPacketSize) + pm.currentPacketExtractionOffset,
 								sizeof(uint16_t) - pm.currentPacketExtractionOffset, 0);
